distinguish the command version from the module version
this means changes in the API of App::Wax, which is effectively
internal, don't trigger major SemVer bumps in wax, which is
the public API. the latter only needs a major bump if the
options are changed in an incompatible way

this is similar to how e.g. ffmpeg distinguishes the version of the
tool (e.g. 3.4.2) from that of its libraries (e.g. libavcodec 57.107.100)